Avoid using double for productivities and tph (#13)
* Avoid using double for productivities and tph

* CR - Use @Value

+++ src/main/java/com/mercadolibre/planning/model/api/domain/entity/forecast/HeadcountProductivity.java
@@ -38,7 +38,7 @@ public class HeadcountProductivity {
     @Enumerated(EnumType.STRING)
     private ProcessName processName;
 
-    private double productivity;
+    private long productivity;
 
     @Enumerated(EnumType.STRING)
     private MetricUnit productivityMetricUnit;
